Restructure packaging for google pay
COAND-497
