+++ googlepay/src/main/java/com/adyen/checkout/googlepay/BillingAddressParameters.kt
@@ -5,7 +5,7 @@
  *
  * Created by caiof on 18/7/2019.
  */
-package com.adyen.checkout.googlepay.model
+package com.adyen.checkout.googlepay
 
 import com.adyen.checkout.core.exception.ModelSerializationException
 import com.adyen.checkout.core.model.ModelObject
