+++ googlepay/src/main/java/com/adyen/checkout/googlepay/GooglePayComponent.kt
@@ -27,6 +27,8 @@ import com.adyen.checkout.components.util.PaymentMethodTypes
 import com.adyen.checkout.core.log.LogUtil
 import com.adyen.checkout.core.log.Logger
 import com.adyen.checkout.googlepay.GooglePayComponent.Companion.PROVIDER
+import com.adyen.checkout.googlepay.internal.provider.GooglePayComponentProvider
+import com.adyen.checkout.googlepay.internal.ui.GooglePayDelegate
 import kotlinx.coroutines.flow.Flow
 
 /**
