+++ googlepay/src/main/java/com/adyen/checkout/googlepay/internal/data/model/GooglePayPaymentMethodModel.kt
@@ -5,7 +5,7 @@
  *
  * Created by caiof on 17/7/2019.
  */
-package com.adyen.checkout.googlepay.model
+package com.adyen.checkout.googlepay.internal.data.model
 
 import com.adyen.checkout.core.exception.ModelSerializationException
 import com.adyen.checkout.core.model.ModelObject
@@ -17,7 +17,7 @@ import org.json.JSONException
 import org.json.JSONObject
 
 @Parcelize
-data class GooglePayPaymentMethodModel(
+internal data class GooglePayPaymentMethodModel(
     var type: String? = null,
     var parameters: CardParameters? = null,
     var tokenizationSpecification: PaymentMethodTokenizationSpecification? = null,
