+++ googlepay/src/main/java/com/adyen/checkout/googlepay/internal/data/model/PaymentMethodTokenizationSpecification.kt
@@ -5,7 +5,7 @@
  *
  * Created by caiof on 17/7/2019.
  */
-package com.adyen.checkout.googlepay.model
+package com.adyen.checkout.googlepay.internal.data.model
 
 import com.adyen.checkout.core.exception.ModelSerializationException
 import com.adyen.checkout.core.model.ModelObject
@@ -16,7 +16,7 @@ import org.json.JSONException
 import org.json.JSONObject
 
 @Parcelize
-data class PaymentMethodTokenizationSpecification(
+internal data class PaymentMethodTokenizationSpecification(
     var type: String? = null,
     var parameters: TokenizationParameters? = null,
 ) : ModelObject() {
